feat: add hooks that consume boardContext
useBoard hook will return every variable that board context will provide

useCard takes an id and returns the card with that id.
it also contains some functions that manipulates the card

useColumn takes an id and returns the column with that id.
it also contains some functions that manipulates the column